requirements.txt
openai>=1.51.0
chainlit>=2.10.0
httpx>=0.27.2
python-dotenv>=1.0.1
python-dateutil>=2.8.0
pytz>=2024.1
